refact(QuickAddEditableView): extract useEditCrossSellOfferForm hook
